refactor(api): update parameter handling in route functions to use Promises for better async handling and improve readability fix(api): update references from params to destructured variables for clarity and consistency across route handlers chore(api): move schema imports from lib to types for better organization and maintainability
feat(cluster-schemas.ts): add Zod schemas for cluster management to ensure data validation and consistency
feat(provider-schemas.ts): introduce Zod schemas for provider management to validate input data and maintain integrity
