+++ adaptive-app/src/types/cluster-schemas.ts
