+++ adaptive-app/src/types/provider-schemas.ts
